Feat: Adding checkers to controller

+++ api/v1/modules/user/controllers/UserController.ts
@@ -155,8 +155,16 @@ export const deleteUser: any = async (req: Request, res: Response) => {
 export const getProfessionalExperienceByUserId: any = async (req: Request, res: Response) => {
   try {
     const userId = req.params.id;
-    const data = await UserService.getProfessionalExperienceByUserId(userId);
-    res.status(200).send(data);
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkGetProfessionalExperienceByUserId(userId,token);
+    if(check==='Profesional Experience not found'){
+      res.status(404).send(check);
+    }else if(check==='No token provided' || check==='Unauthorized'){
+      res.status(401).send(check);
+    }else{
+      const data = await UserService.getProfessionalExperienceByUserId(userId);
+      res.status(200).send(data);
+    }
   } catch (error: any) {
     console.error(error);
     res.status(500).send(error.message);
@@ -165,8 +173,16 @@ export const getProfessionalExperienceByUserId: any = async (req: Request, res: 
 
 export const createProfessionalExperience: any = async (req: Request, res: Response) => {
   try {
-    const data = await UserService.createProfessionalExperience(req.body);
-    res.status(200).send(data);
+    const token = req.headers.authorization ?? '';
+    const check= await UserMiddleware.checkCreateProfessionalExperience(req.body,token)
+    if(check==='Invalid candidate' || check==='Missing required fields' || check==='No data to update'){
+      res.status(400).send(check);
+    }else if (check === 'Unauthorized' || check === 'No token provided') {
+      res.status(401).send(check);
+    }else{
+      const data = await UserService.createProfessionalExperience(req.body);
+      res.status(200).send(data);
+    }
   } catch (error: any) {
     console.error(error);
     res.status(500).send(error.message);
@@ -176,8 +192,18 @@ export const createProfessionalExperience: any = async (req: Request, res: Respo
 export const updateProfessionalExperience: any = async (req: Request, res: Response) => {
   try {
     const id = req.params.id;
-    const data = await UserService.updateProfessionalExperience(id, req.body);
-    res.status(200).send(data);
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkUpdateProfessionalExperience(id, token,req.body);
+    if (check === 'Professional experience not found') {
+      res.status(404).send(check);
+    }else if (check === 'No data to update' || check==='Missing required fields') {
+      res.status(400).send(check);
+    } else if (check === 'Unauthorized' || check === 'No token provided') {
+      res.status(401).send(check);
+    } else {
+      const data = await UserService.updateProfessionalExperience(id, req.body);
+      res.status(200).send(data);
+    }
   } catch (error: any) {
     console.error(error);
     res.status(500).send(error.message);
@@ -187,8 +213,16 @@ export const updateProfessionalExperience: any = async (req: Request, res: Respo
 export const deleteProfessionalExperience: any = async (req: Request, res: Response) => {
   try {
     const id = req.params.id;
-    const data = await UserService.deleteProfessionalExperience(id);
-    res.status(200).send(data);
+    const token = req.headers.authorization ?? '';
+    const check = await UserMiddleware.checkDeleteProfessionalExperience(id, token);
+    if (check === 'Professional experience not found') {
+      res.status(404).send(check);
+    } else if (check === 'Unauthorized' || check === 'No token provided') {
+      res.status(401).send(check);
+    } else {
+      const data = await UserService.deleteProfessionalExperience(id);
+      res.status(200).send(data);
+    }
   } catch (error: any) {
     console.error(error);
     res.status(500).send(error.message);
